Began refactoring the staking contract to be a subtree.
